Refactor code using packages
