+++ src/main/java/duke/DukeException.java
@@ -1,9 +1,11 @@
+package duke;
+
 /**
- * Represents exceptions specific to Duke.
+ * Represents exceptions specific to duke.Duke.
  */
 public class DukeException extends Exception {
     /**
-     * Constructor for a DukeException instance.
+     * Constructor for a duke.DukeException instance.
      * @param errorMessage Error message to be displayed.
      */
     public DukeException(String errorMessage) {
