+++ src/main/java/duke/command/TaskCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.task.Task;
+
 public class TaskCommand extends Command {
     private final Task newTask;
